Move tests to separate submodule
